requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.4
filelock==3.13.4
fsspec==2024.3.1
idna==3.7
Jinja2==3.1.3
llvmlite==0.42.0
MarkupSafe==2.1.5
more-itertools==10.2.0
mpmath==1.3.0
networkx==3.3
numba==0.59.1
numpy==1.26.4
openai-whisper==20231117
pytube==15.0.0
regex==2024.4.16
requests==2.31.0
six==1.16.0
sqlparse==0.5.0
sympy==1.12
tiktoken==0.6.0
torch==2.2.2
tqdm==4.66.2
typing_extensions==4.11.0
urllib3==2.2.1
